pyproject.toml
[project]
name = "auto-analytics"
version = "0.1.0"
description = "Auto Analytics AI Agent System - データ分析の民主化を実現するマルチエージェントシステム"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.13",
    "asyncpg>=0.29.0",
    "google-generativeai>=0.8.5",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.17.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "structlog>=25.4.0",
    "google-adk>=1.4.1",
    "mcp>=1.9.4",
    "jinja2>=3.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "markdown>=3.8.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "jupyter>=1.1.1",
    "mypy>=1.16.1",
    "notebook>=7.4.3",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
]
